backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydub==0.25.1
scipy==1.11.4
nltk==3.8.1
textstat==0.7.3

